GROUPBY on StokcItem
Template : Vouchers-Sales-Purchase-With-StockItems-Advanced.xls
Description : The feature facilitates grouping on the StockItem. i.e. if
there are multiple items in a Sales or Purchase Invoice and if the items
are repeated, then this feature allows grouping on the StockItem.

Example :
An invoice contains :
Item A   Qty 2
Item B   Qty 5
Item A   Qty 3
Item C   Qty 10
Item A   Qty 5
Item B   Qty 2

The Grouping feature groups it as follows :

Item A   Qty 10
Item B   Qty 7
Item C   Qty 10