v11 r3.7
Changes in Vouchers-V11-Bank-statement.xls

Changes have been done to prevent duplicate vouchers (if same sheet is
imported again).

Notes:
a. If the same data/sheet is imported twice, then the entries will be duplicated in Tally.
(This is because there is no unique identification number is present in this sheet).

b. To overcome this problem, we now use the VoucherNo as unique-id. i.e if VoucherNo is given,
   then, unique-id will be taken as :
   unique-id =  voucher-no + voucher-date + voucher-type

IMP : Howevever, you must ensure that you do NOT repeat the same voucher-no for
another voucher-type during a given financial year.