v11 r4.3
Template: Vouchers-V11-Sales-Purchase-With-Stockitems-Advanced.xls

Issue
TCS amount rounded-off to 0 decimals. This was kept purposefully
as in the older releases of Tally, the TCS amount was also auto-rounded
off to Integer value.

Solution
In the latest release of Tally Prime (Feb-2023), the TCS amount is accepted in decimals
also. Hence, we have also corrected this in the udiMagic template.