v12 r1.2 SPIA
Fixes: PLACEOFSUPPLY

PLACEOFSUPPLY is not taken from Column M. If it is empty, then StateName
is taken from Column K as the place of supply.

This works in the following cases:
Sales - Local, Inter-State
Purchase - Local

Note:
It does not work in case of Purchase Inter-state if Column M is empty.
i.e. Column M (PlaceOfSupply is Mandatory in case of Inter-state Purchases)