Changes in SPIA xml
When there are no items, the grouping of Sales/Purchase ledger is done
based on:
Ledger name (CB) + Tax (DI or DK)

Previously, the grouping was done differently:
a. If CB is not empty (i.e. ledger name is specified), then group on
Ledger-names
b. If CB is empty, then group on DI (IGST rate) or DK (CGST rate)